moved metrics_plots into metrics module, removed whitespace and added empty line at end of file
