+++ fairlearn/metrics/_plots/__init__.py
@@ -1,0 +1,17 @@
+# Copyright (c) Microsoft Corporation and Fairlearn contributors.
+# Licensed under the MIT License.
+
+"""Submodule for visualizing metrics.
+
+This submodule contains functions for visualizing metrics generated by the
+fairlearn.metrics module.
+"""
+
+
+from ._plot_disparities_in_performance import plot_disparities_in_performance
+from ._plot_disparities_in_metric import plot_disparities_in_metric
+
+__all__ = [
+    "plot_disparities_in_performance",
+    "plot_disparities_in_metric"
+]
