+++ fairlearn/metrics/_plots/_plot_disparities_in_metric.py
